adding result saving to txt file and dork list from file
adding an option to import dork list from a file in the google mode by typing the dork like \f: file_name
file should be in the main katana dir.
adding the option to save the result to a text file in the main dir. by ending the dork, or file name with SPACE+\s